miracle part 1
git-svn-id: https://mlt.svn.sourceforge.net/svnroot/mlt/trunk/mlt@39 d19143bc-622f-0410-bfdd-b5b2a6649095

+++ mlt/src/valerie/valerie.c
@@ -311,9 +311,9 @@ valerie_error_code valerie_unit_play( valerie this, int unit )
 /** Play the unit at specified speed.
 */
 
-valerie_error_code valerie_unit_play_at_speed( valerie this, int unit, double speed )
+valerie_error_code valerie_unit_play_at_speed( valerie this, int unit, int speed )
 {
-	return valerie_execute( this, 10240, "PLAY U%d %e", unit, speed );
+	return valerie_execute( this, 10240, "PLAY U%d %d", unit, speed );
 }
 
 /** Stop playback on the specified unit.
